creating separate class for MinkowskiNeighbors
also adding a more robust example of using a custom distance function
