+++ river/test_estimators.py
@@ -77,8 +77,9 @@ def iter_estimators_which_can_be_tested():
         imblearn.RandomSampler,
         model_selection.SuccessiveHalvingClassifier,
         neighbors.BaseKNN,
-        neighbors.NearestNeighbors,
         neighbors.KNNRegressor,
+        neighbors.MinkowskiNeighbors,
+        neighbors.NearestNeighbors,
         neural_net.MLPRegressor,
         preprocessing.PreviousImputer,
         preprocessing.OneHotEncoder,
